feat: Update GetRecepcionProductoResponse and related classes to use BigDecimal for pricing and enhance response mapping

+++ src/main/java/bo/com/micrium/modulobase/modulos/compra/controllers/dtos/recepcion/get/GetDetalleRecepcionResponse.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Getter
@@ -20,6 +21,6 @@ public class GetDetalleRecepcionResponse implements Serializable {
     private Date fechaVencimiento;
     private String lote;
     private Integer cantidad;
-    private Double precio;
+    private BigDecimal precio;
 }
